# internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/namespace:
        post:
            tags:
                - Namespace
            operationId: Namespace_CreateNamespace
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.CreateNamespaceRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.CreateNamespaceReply'
    /v1/namespace/{uid}:
        get:
            tags:
                - Namespace
            operationId: Namespace_GetNamespace
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.NamespaceItem'
        put:
            tags:
                - Namespace
            operationId: Namespace_UpdateNamespace
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.UpdateNamespaceRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.UpdateNamespaceReply'
        delete:
            tags:
                - Namespace
            operationId: Namespace_DeleteNamespace
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.DeleteNamespaceReply'
    /v1/namespace/{uid}/status:
        put:
            tags:
                - Namespace
            operationId: Namespace_UpdateNamespaceStatus
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.UpdateNamespaceStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.UpdateNamespaceStatusReply'
    /v1/namespaces:
        get:
            tags:
                - Namespace
            operationId: Namespace_ListNamespace
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.ListNamespaceReply'
    /v1/namespaces/select:
        get:
            tags:
                - Namespace
            operationId: Namespace_SelectNamespace
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lastUID
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.SelectNamespaceReply'
...
components:
    schemas:
        goddess.api.v1.CreateNamespaceReply:
            type: object
            properties: {}
        goddess.api.v1.CreateNamespaceRequest:
            type: object
            properties:
                name:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
        goddess.api.v1.DeleteNamespaceReply:
            type: object
            properties: {}
...
        goddess.api.v1.ListNamespaceReply:
            type: object
            properties:
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/goddess.api.v1.NamespaceItem'
...
        goddess.api.v1.NamespaceItem:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
                status:
                    type: integer
                    format: enum
        goddess.api.v1.NamespaceItemSelect:
            type: object
            properties:
                value:
                    type: string
                label:
                    type: string
                disabled:
                    type: boolean
                tooltip:
                    type: string
...
        goddess.api.v1.SelectNamespaceReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/goddess.api.v1.NamespaceItemSelect'
                total:
                    type: string
                lastUID:
                    type: string
                hasMore:
                    type: boolean
...
        goddess.api.v1.UpdateNamespaceReply:
            type: object
            properties: {}
        goddess.api.v1.UpdateNamespaceRequest:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
        goddess.api.v1.UpdateNamespaceStatusReply:
            type: object
            properties: {}
        goddess.api.v1.UpdateNamespaceStatusRequest:
            type: object
            properties:
                uid:
                    type: string
                status:
                    type: integer
                    format: enum
...